backend/requirements.txt
# promptos_backend/requirements.txt
fastapi
uvicorn
loguru
pydantic
redis
motor
python-multipart
python-dotenv
GitPython
google-generativeai
httpx
